+++ src/main.js
@@ -1,11 +1,13 @@
 import showFirst from './vistas/inicio.js'
 import showLogin from './vistas/registro.js'
 import publications from './vistas/publicaciones.js'
+import {registry, observer, closeSession} from './firebase/firebaseAuth.js'
+import createPost from './firebase/firestore.js'
+
 
 window.addEventListener('hashchange', () => {
     router(window.location.hash);
 });
-
 
 window.addEventListener('load', () => {
     router('#/')
@@ -14,7 +16,7 @@ window.addEventListener('load', () => {
 
 let content = document.getElementById("container"); 
 
-const router = (route) => {
+export const router = (route) => {
     content.innerHTML = '';
     switch(route) {
         case '#/':
@@ -42,149 +44,18 @@ const router = (route) => {
     }
 }
 
-const registry= () => {
-// Firebase de registro 
-const registro= document.querySelector('#registro');
-registro.addEventListener ('submit', (e) => {
-    e.preventDefault();
-    const email = document.querySelector('#input-correo').value;
-    const pws1= document.querySelector('#input-psw1').value;
-    const pws2= document.querySelector('#input-psw2').value;
-    
-    firebase.auth().createUserWithEmailAndPassword(email, pws1)
-    .then (() => {
-        verify();
-    })
-    .catch((error) => {
-    let errorCode = error.code;
-    let errorMessage = error.message;
-    alert(errorMessage);
-    })
-});
-
-// Firebase de inicio de sesion
-
-const signInForm= document.querySelector('#Login');
-signInForm.addEventListener('submit', (e) => {
-    e.preventDefault();
-    const loginEmail = document.querySelector('#login-email').value;
-    const loginPws= document.querySelector('#login-psw').value;
-    //console.log(loginEmail,loginPws);
-    firebase.auth().signInWithEmailAndPassword(loginEmail, loginPws)
-    .then ((result) => {
-        
-        let emailVerified = result.user.emailVerified;
-        
-        if(emailVerified){
-            router('#/Home');
-        }
-        else{
-            console.log('Debes verificar tu correo');
-        }
-    })
-    .catch((error) => {
-    let errorCode = error.code;
-    let errorMessageSign = error.message;
-    alert(errorMessageSign);
-    })
-});
-}
-
-const observer = () => {
-firebase.auth().onAuthStateChanged((user) => {
-    if (user) {
-        console.log('Existe Usuario Activo');
-        // User is signed in.
-    
-    let displayName = user.displayName;
-    let email = user.email;
-    
-    console.log('########');
-    console.log(user.emailVerified);
-    console.log('########');
-
-    let emailVerified = user.emailVerified;
-    let photoURL = user.photoURL;
-    let isAnonymous = user.isAnonymous;
-    let uid = user.uid;
-    let providerData = user.providerData;
-    // ...
-    } 
-    else {
-      // User is signed out.
-        console.log('vuelva a iniciar sesión');
-    }
-});
-}
-
 const modalInicio = () => {
 
-let modal = document.querySelector(".modal");
-let btn = document.getElementById("btninicio-sesion");
-let span = document.getElementsByClassName("close")[0];
-
-btn.onclick =  () => {
-    modal.style.visibility= "visible";
-    registro.style.visibility= "hidden";
-}
-span.onclick =  () => {
-    modal.style.visibility = "hidden";
-    registro.style.visibility= "visible";
-}
-}
-
-//Cerrar Sesión
-const closeSession = () => {
-    firebase.auth().signOut()
-    .then(() => {
-        console.log('Saliendo...');
-    }) 
-    .catch((error) => {
-        console.log(error);
-    })
-}
-
-//Verificación de usuario
-const verify = () => {
-let user = firebase.auth().currentUser;
-user.sendEmailVerification().then(() => {
-    console.log('Enviando correo...');
-    // Email sent.
-}).catch((error) => {
-    // An error happened.
-    console.log(error);
-});
-}
-
-/*const createPost = () => {
-const post = document.querySelector('#form-post');
-post.addEventListener('submit', (e) => {
-    e.preventDefault();
+    let modal = document.querySelector(".modal");
+    let btn = document.getElementById("btninicio-sesion");
+    let span = document.getElementsByClassName("close")[0];
     
-const title = document.getElementById('title-post');
-const description = document.getElementById('description');
-console.log(title.value, description.value);
-})
-}*/
-
-const createPost = () => {
-    let publicar = document.querySelector(".publicar");
-    publicar.addEventListener('submit', (e) => {
-        e.preventDefault();
-        
-        let title = document.getElementById('title-post').value;
-        let description = document.getElementById('description').value;
-
-        //Inicialize Cloud Firestore throught Firebase
-        firebase.firestore().collection("publications").add({
-            Titulo: title,
-            descripcion: description
-        })
-
-        .then(function(docRef) {
-            console.log("Document written with ID: ", docRef.id);
-        })
-        .catch(function(error) {console.error("Error adding document: ", error);
-        });
-    })
-};
+    btn.onclick =  () => {
+        modal.style.visibility= "visible";
+        registro.style.visibility= "hidden";
+    }
+    span.onclick =  () => {
+        modal.style.visibility = "hidden";
+        registro.style.visibility= "visible";
+    }
+    }
